+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 0.2.1
 
 * Fix issue with MOPE in case it runs as an external dependency.
+* Add `LimitedSet::contains(other)` which performs contains-all functionality (not part of STL).
 
 # 0.2